Update return value of split_string from int to size_t

+++ include/orcania.h
@@ -184,7 +184,7 @@ char * mstrcatf(char * source, const char * message, ...);
  * return_array must be free'd after use
  * you can use free_string_array to free return_array
  */
-int split_string(const char * string, const char * separator, char *** return_array);
+size_t split_string(const char * string, const char * separator, char *** return_array);
 
 /**
  * Clean an array of strings
